requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
reportlab
python-dotenv